+++ server/helpers/custom-validators/video-abuses.ts
@@ -1,5 +1,5 @@
 import { Response } from 'express'
-import * as validator from 'validator'
+import validator from 'validator'
 import { CONSTRAINTS_FIELDS, VIDEO_ABUSE_STATES } from '../../initializers/constants'
 import { exists } from './misc'
 import { VideoAbuseModel } from '../../models/video/video-abuse'